refactor(sync): split versioned map tests for GH727
